% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{add_user_group}
\alias{add_user_group}
\title{Add a new user to a given group}
\usage{
add_user_group(group_id, user_id)
}
\arguments{
\item{group_id}{the group id, see \code{list_groups()}}

\item{user_id}{the user account id, see \code{list_users()}}
}
\value{
A status code
}
\description{
Add a new user to a given group
}
\examples{
\dontrun{add_user_group("GR_1234567890AbCdE", "UR_1234567890AbCdE")}
}
