% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{bulk-download}
\alias{bulk-download}
\alias{request_downloads}
\alias{download_requested}
\alias{download_requested.qualtrics_download}
\title{Request downloads}
\usage{
request_downloads(surveyIds)

download_requested(requests, format = "csv", saveDir = NULL,
  verbose = FALSE, ...)

\method{download_requested}{qualtrics_download}(requests, format = "csv",
  saveDir = NULL, verbose = FALSE, ...)
}
\arguments{
\item{surveyIds}{A vector of survey ids.}

\item{requests}{Requests, as returned by \code{request_downloads}}

\item{format}{file format json, by default (can be csv or tsv). We don't provide SPSS yet.}

\item{saveDir}{path to a local directory to save the exported file. Default is a temporary file in rds format that is removed at the end of your session. Default name will be the surveyId.}

\item{verbose}{default FALSE}

\item{...}{Any other option.}
}
\value{
A \code{request_downloads} returns object of class \code{qualtrics_download} while
\code{download_requested} returns a \code{list}.
}
\description{
Request survey response downloads.
}
\details{
The \href{https://www.qualtrics.com}{Qualtrics} API takes some time to prepare a downloads, hence
\code{\link{download_requested}} taking a some time to download the requested data. What actually happens
is that first a download is \emph{requested}, after the request Qualtrics prepares a file that
\code{\link{download_requested}} downloads when it is ready. It is therefore more efficient,
when the user wants to download data from multiple surveys, more efficient to request each and every
download before checking whether they are ready. Those performances will headvily depend on the number
of surveys and the amount of data the ultimately user wants to obtain.
}
\section{Functions}{

\itemize{
  \item{\code{request_downloads}: Request download (prepares files).}
  \item{\code{download_requested}: Download requested files.}
}
}

\examples{
(requests <- request_downloads(c(1,2)))
data <- download_requested(requests)

}
