% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{copy_survey}
\alias{copy_survey}
\title{Create a copy of a survey}
\usage{
copy_survey(survey_id, copy_name, user_id = NULL)
}
\arguments{
\item{survey_id}{the survey id}

\item{copy_name}{the name of the survey copy}

\item{user_id}{if provided, copies the survey to another account}
}
\value{
The id of the copied survey
}
\description{
The survey can be copied within the same account or to another user account.
Default uses the user id returned by \code{\link{who_am_i}} function. When a survey gets
copied it will appear in the 'Uncategorized' folder
}
\examples{
\dontrun{copy_survey("SV_012345678901234", "survey_copy")}
\dontrun{copy_survey("SV_012345678901234", "survey_copy", "UR_012345678912345")}
}
