% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{create_block}
\alias{create_block}
\title{Create a new block}
\usage{
create_block(survey_id, type = "Standard", description = "Block", ...)
}
\arguments{
\item{survey_id}{the survey id}

\item{type}{block type (Standard, Default or Trash)}

\item{description}{block description}

\item{...}{block options}
}
\value{
A list with the associated block id and flow id
}
\description{
To retrieve a specific survey block, you need to pass the survey id and the block id. The latter
can be retrieved using the get_survey_flow function or the get_survey function.
}
\examples{
\dontrun{
create_block(
"SV_012345678901234", 
"Standard",
"New Block",
"BlockLocking": "false",
"RandomizeQuestions": "false",
"BlockVisibility": "Collapsed"
)
}
}
