% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{create_contact}
\alias{create_contact}
\title{Create a new contact in a mailing list}
\usage{
create_contact(mailinglist_id, contact_options)
}
\arguments{
\item{mailinglist_id}{ID of the mailing list in which to create the contact}

\item{contact_options}{list of new contact information}
}
\value{
The id of the created mailing list
}
\description{
create a new contact in a mailing list. In addition to the mailing
list id, you can pass a named list of parameters to define your contact. All
parameters are optional but empty parameters cannot be passed. Parameters are:
firstName, lastName, email, externalDataRef, language, unsubscribed and embeddedData. The
embeddedData element must a string representation of a JSON object.
}
\examples{
\dontrun{
new_contact <- list(
"firstName" = "James",
"lastName" = "Paddignton",
"email" = "james.pad@gmailing.com")
new_contact_id <- create_contact("ML_0HT4q2Ni634kLhH", new_contact)}
}
