% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{create_distribution}
\alias{create_distribution}
\title{Create a new email distribution}
\usage{
create_distribution(
  survey_id,
  library_id,
  message_id,
  mailing_list_id,
  from_name,
  reply_to_email,
  from_email,
  subject,
  type,
  send_date
)
}
\arguments{
\item{survey_id}{the id of survey to copy}

\item{library_id}{Library ID of the message}

\item{message_id}{id of the message item}

\item{mailing_list_id}{id of the contact list for the distribution}

\item{from_name}{appearing sender name, default is Qualtrics}

\item{reply_to_email}{reply email, default is \email{noreply@qualtrics.com}}

\item{from_email}{sender email, default is \email{noreply@qualtrics.com}}

\item{subject}{email subject, default is "Survey Distribution"}

\item{type}{Allowed values: Individual, Multiple, Anonymous}

\item{send_date}{date for distribution to be sent}
}
\value{
The created distribution id
}
\description{
Create a new email distribution
}
