% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{create_division}
\alias{create_division}
\title{Create a new division}
\usage{
create_division(name, division_admins = NULL, permissions = NULL)
}
\arguments{
\item{name}{group id}

\item{division_admins}{A list of of user ids to make division admins}

\item{permissions}{A list defining the division permissions.}
}
\value{
The id of the created division
}
\description{
When setting permissions, any permissions left undefined will be enabled by
default. If no permissions are included at all, then a default set of
predefined permissions will be used. If a division permission is disabled at
the organization level, then it will be disabled at the division level
regardless of its state.
}
\examples{
\dontrun{create_division("DV_1234567890AbCdE")}
}
