% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-data-privacy.R
\name{create_erasing_request}
\alias{create_erasing_request}
\title{create_user_api_token creates a users api token}
\usage{
create_erasing_request(list_emails, search_only = TRUE)
}
\arguments{
\item{list_emails}{A vector of emails.}

\item{search_only}{Set to \code{TRUE} to perform a dryrun (default).}
}
\value{
A \code{data.frame}.
}
\description{
create_user_api_token creates a users api token
}
\examples{
\dontrun{create_erasing_request("ddd@hotmail.com")}
\dontrun{create_erasing_request(c("person1@theemail.com", "person1@theemail.com"),TRUE)}
}
