% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{create_mailinglist}
\alias{create_mailinglist}
\title{Create a new mailing list}
\usage{
create_mailinglist(library_id, name, category = NULL)
}
\arguments{
\item{library_id}{the library_id}

\item{name}{Name of the new mailing list}

\item{category}{Category in which to create the new mailing list}
}
\value{
The id of the created mailing list
}
\description{
Create a new mailing list
}
\examples{
\dontrun{create_mailinglist("UR_1234567890AbCdE","New Emails")}
}
