% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{create_message}
\alias{create_message}
\title{create_message creates a message}
\usage{
create_message(description, messages, category, library_id)
}
\arguments{
\item{description}{defines a description for the message}

\item{messages}{A JSON object - mapping language codes to messages}

\item{category}{invite, inactiveSurvey, reminder, thankYou, endOfSurvey, general, validation, lookAndFeel, emailSubject, smsInvite}

\item{library_id}{The Library ID can be one of the following IDs: User ID, Group ID, ThreeSixty ID}
}
\value{
id of created message
}
\description{
create_message creates a message
}
