% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{create_reminder_distribution}
\alias{create_reminder_distribution}
\title{Create a reminder distribution}
\usage{
create_reminder_distribution(
  parent_distribution_id,
  from_email = "noreply@qualtrics.com",
  from_name = "Qualtrics",
  reply_to_email = "noreply@qualtrics.com",
  subject,
  message_id,
  library_id,
  send_date
)
}
\arguments{
\item{parent_distribution_id}{id of the library item}

\item{from_email}{sender email, default is \email{noreply@qualtrics.com}}

\item{from_name}{appearing sender name, default is Qualtrics}

\item{reply_to_email}{reply email, default is \email{noreply@qualtrics.com}}

\item{subject}{email subject, default is "Survey Distribution"}

\item{message_id}{id of the message item}

\item{library_id}{Library ID of the message}

\item{send_date}{date for distribution to be sent}
}
\value{
The reminder distribution id
}
\description{
Based on an existing email distribution, create a reminder to 
take a survey for recipients with non completed surveys. 
Recipients who have opted out will not receive the reminder.
}
\details{
The sender email parameter can only be sent from another email if
the organization authorizes a send on its behalf.
}
