% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-survey.R
\name{create_survey}
\alias{create_survey}
\title{Create a survey}
\usage{
create_survey(survey_name, language = "EN", project_category = "CORE")
}
\arguments{
\item{survey_name}{the survey name}

\item{language}{creation language, default is EN}

\item{project_category}{default is CORE (can be CX, EX, BX, PX)}
}
\value{
A execution status code
}
\description{
Create a survey
}
\examples{
\dontrun{delete_survey("SV_012345678901234")}
}
