% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{create_user}
\alias{create_user}
\title{Create a new user account}
\usage{
create_user(...)
}
\arguments{
\item{...}{a named list of account parameters to update}
}
\value{
The created user id
}
\description{
Create a new user account
}
\section{Account parameters}{

user settings that can be possed are user_name, first_name, last_name,
user_type, language, time_zone, division_id, account_expiration_date,
status and permissions
}

\examples{
\dontrun{create_user(c("status" = "Enabled", "language" = "EN"))}
}
