% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-users.R
\name{create_user_api_token}
\alias{create_user_api_token}
\title{Create an api token for a user account}
\usage{
create_user_api_token(user_id)
}
\arguments{
\item{user_id}{the user id}
}
\value{
An new api token value
}
\description{
Create an api token for a user account
}
\examples{
\dontrun{create_user_api_token("UR_012345678912345")}
}
