% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{delete_block}
\alias{delete_block}
\title{Delete a block}
\usage{
delete_block(survey_id, block_id)
}
\arguments{
\item{survey_id}{the survey id}

\item{block_id}{the block id}
}
\value{
An execution status code
}
\description{
Delete a block
}
\examples{
\dontrun{
delete_block("SV_012345678901234", "BL_723z35LY23KCZ4p")
}
}
