% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{delete_contact}
\alias{delete_contact}
\title{Delete a contact from a mailing list}
\usage{
delete_contact(mailinglist_id, contact_id)
}
\arguments{
\item{mailinglist_id}{the id of mailing list to delete}

\item{contact_id}{ID of the contact to update (see list_contacts)}
}
\value{
A status code
}
\description{
Delete a contact from a mailing list
}
\examples{
\dontrun{delete_contact("ML_1234567890AbCdE","MLRP_eaI7CCX9bxAbRt3")}
}
