% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{delete_distribution}
\alias{delete_distribution}
\title{Delete a survey distribution}
\usage{
delete_distribution(distribution_id)
}
\arguments{
\item{distribution_id}{the distribution id}
}
\value{
A status execution code
}
\description{
Delete a survey distribution
}
\examples{
\dontrun{delete_distribution("EMD_012345678901234")}
}
