% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic-library.R
\name{delete_graphic}
\alias{delete_graphic}
\title{delete_graphic retrieves a full survey object}
\usage{
delete_graphic(libraryId, graphicId)
}
\arguments{
\item{libraryId}{the distribution id}

\item{graphicId}{the distribution id}
}
\value{
A status code
}
\description{
delete_graphic retrieves a full survey object
}
\examples{
\dontrun{delete_graphic("UR_6xMwYO5zW4FklUh", "IM_6s1QFALhGhB8zVr")}
}
