% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-groups.R
\name{delete_group}
\alias{delete_group}
\title{Delete a group}
\usage{
delete_group(group_id)
}
\arguments{
\item{group_id}{id of the group to delete}
}
\value{
A status execution code
}
\description{
Delete a group
}
\examples{
\dontrun{delete_group("GR_1234567890AbCdE")}
}
