% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{delete_mailinglist}
\alias{delete_mailinglist}
\title{Delete an existing mailing list}
\usage{
delete_mailinglist(mailinglist_id)
}
\arguments{
\item{mailinglist_id}{the id of mailing list to delete}
}
\value{
A status code
}
\description{
Delete an existing mailing list
}
\examples{
\dontrun{delete_mailinglist("ML_1234567890AbCdE")}
}
