% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{delete_message}
\alias{delete_message}
\title{delete_message creates a message}
\usage{
delete_message(library_id, message_id)
}
\arguments{
\item{library_id}{The Library ID can be one of the following IDs: User ID, Group ID, ThreeSixty ID}

\item{message_id}{Message Id run list_messages}
}
\value{
A \code{tibble}.
}
\description{
delete_message creates a message
}
