% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{delete_response}
\alias{delete_response}
\title{Delete a survey response}
\usage{
delete_response(surveyId, responseId, decrementQuotas)
}
\arguments{
\item{surveyId}{ID of the survey to delete the response from}

\item{responseId}{ID of the response to be deleted}

\item{decrementQuotas}{If true, any relevant quotas will be decremented}
}
\value{
A \code{status}.
}
\description{
Delete a survey response
}
\examples{
\dontrun{delete_response("SV_012345678912345", "R_012345678912345", TRUE)}
}
