% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{delete_sms_distribution}
\alias{delete_sms_distribution}
\title{Delete an sms survey distribution}
\usage{
delete_sms_distribution(smsdistribution_id)
}
\arguments{
\item{smsdistribution_id}{the distribution id}
}
\value{
A status execution code
}
\description{
Delete an sms survey distribution
}
\examples{
\dontrun{delete_sms_distribution("SMSD_012345678901234")}
}
