% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-survey.R
\name{delete_survey}
\alias{delete_survey}
\title{Delete a survey}
\usage{
delete_survey(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A execution status code
}
\description{
Delete a survey
}
\examples{
\dontrun{delete_survey("SV_012345678901234")}
}
