% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-questions.R
\name{delete_survey_question}
\alias{delete_survey_question}
\title{Delete a survey question}
\usage{
delete_survey_question(survey_id, question_id)
}
\arguments{
\item{survey_id}{the survey id}

\item{question_id}{the question id (see list_survey_questions)}
}
\value{
A list
}
\description{
Delete a survey question
}
