% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user account}
\usage{
delete_user(user_id)
}
\arguments{
\item{user_id}{the user account id to delete}
}
\value{
A status execution code
}
\description{
Delete a user account
}
\examples{
\dontrun{delete_user("UR_012345678912345")}
}
