% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.construct_oauth}
\alias{.construct_oauth}
\title{Construct Token Object
Constructs and object for print method}
\usage{
.construct_oauth(token)
}
\arguments{
\item{token}{A list.}
}
\description{
Construct Token Object
Constructs and object for print method
}
\keyword{internal}
