% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_token}
\alias{.get_token}
\title{Get Token
Gets token from options.}
\usage{
.get_token()
}
\description{
Get Token
Gets token from options.
}
\details{
If \code{timeout} options is found then assesses whether refresh is required.
Note that we remove one minute from the current time to allow for small time differences between
token being fetched and call being made.
}
\keyword{internal}
