% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.encrypt}
\alias{.encrypt}
\alias{.decrypt}
\title{Encrypt and Decrypt token
Encrypt and decrypt for security, not saving plain credentials to file.}
\usage{
.encrypt(token)

.decrypt(token)
}
\arguments{
\item{token}{A list.}
}
\description{
Encrypt and Decrypt token
Encrypt and decrypt for security, not saving plain credentials to file.
}
\keyword{internal}
