% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{generate_block_id}
\alias{generate_block_id}
\title{Generate a compatible block id}
\usage{
generate_block_id(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A string of the form BL_FakeBlockID1234
}
\description{
When updating a block (update_block) or creating a new block (create_block), it can be convenient to control the creation
of the block id locally. This convenience function returns a block id that is unique and does not overlap with
existing block ids in the survey.
}
\examples{
\dontrun{
generate_block_id(
"SV_012345678901234", 
)
}
}
