% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{generate_distribution_links}
\alias{generate_distribution_links}
\title{Generate survey distribution links}
\usage{
generate_distribution_links(
  survey_id,
  mailinglist_id,
  description = "Generated Links",
  linktype = "Individual",
  action = "CreateDistribution",
  expirationdate = NULL
)
}
\arguments{
\item{survey_id}{the distribution id}

\item{mailinglist_id}{the distribution id}

\item{description}{the distribution id}

\item{linktype}{the id of survey}

\item{action}{default is CreateDistribution. To generate transaction distribution links, value must be "CreateTransactionBatchDistribution"}

\item{expirationdate}{the id of survey}
}
\value{
The id of the created distribution
}
\description{
Create a distribution without sending any emails. The created distribution
will be of type "GeneratedInvite". The survey must be active before you can
generate a distribution invite. Refer to the documentation below for further
information regarding usage of this API's result.
}
\details{
To retrieve the actual links, use the 'list_distribution_links' call.
}
\examples{
\dontrun{
 generate_distribution_links(
   "SV_erkBAsHrvoJyeYB",
   "ML_7aoriSKinHh8MfP", 
   "Survey Respondents Generated Links"
}

}
