% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{generate_distributionlink}
\alias{generate_distributionlink}
\title{Generate a distribution link}
\usage{
generate_distributionlink(action, survey_id, mailinglist_id, description,
  expirationdate, linktype)
}
\arguments{
\item{action}{To generate transaction distribution links, value must be "CreateTransactionBatchDistribution"}

\item{survey_id}{the distribution id}

\item{mailinglist_id}{the distribution id}

\item{description}{the distribution id}

\item{expirationdate}{the id of survey}

\item{linktype}{the id of survey}
}
\value{
A \code{list}.
}
\description{
Create a distribution without sending any emails. The created distribution
will be of type "GeneratedInvite". The survey must be active before you can
generate a distribution invite. Refer to the documentation below for further
information regarding usage of this API's result.
}
