% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-events.R
\name{get_activity_type}
\alias{get_activity_type}
\title{Retrieve the type of logging (activity) events}
\usage{
get_activity_type()
}
\value{
A \code{list} with the mailing list information
}
\description{
Retrieve the type of logging (activity) events
}
\examples{
\dontrun{get_activity_type()}
}
