% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{get_contact}
\alias{get_contact}
\title{Retrieve a the information related to a certain contact}
\usage{
get_contact(mailinglist_id, contact_id)
}
\arguments{
\item{mailinglist_id}{the mailing list id}

\item{contact_id}{the contact id}
}
\value{
A \code{list} with the mailing list information
}
\description{
This call allows to retrieve all the information regarding a certain contact id. The call returns an
email history for the contact
}
\examples{
\dontrun{get_contact("ML_0HT4q2Ni634kLhH", "MLRP_bwwuYtDMPdK2Scl")}
}
