% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_distribution}
\alias{get_distribution}
\title{Retrieve a given survey distribution object}
\usage{
get_distribution(distribution_id, survey_id)
}
\arguments{
\item{distribution_id}{the distribution id}

\item{survey_id}{the id of the associated survey}
}
\value{
A \code{list}.
}
\description{
Retrieve a given survey distribution object
}
