% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{get_division}
\alias{get_division}
\title{Returns general information about a division}
\usage{
get_division(division_id)
}
\arguments{
\item{division_id}{the division id}
}
\value{
A \code{list} of division definition parameters
}
\description{
Returns general information about a division
}
\examples{
\dontrun{get_division("DV_1234567890AbCdE")}

}
