% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-data-privacy.R
\name{get_erasing_request}
\alias{get_erasing_request}
\title{get_erasing_request lists all erasing requests}
\usage{
get_erasing_request(request_id)
}
\arguments{
\item{request_id}{the id of the past erasing request}
}
\value{
A \code{list}.
}
\description{
get_erasing_request lists all erasing requests
}
\examples{
\dontrun{get_erasing_request("request_id")}
}
