% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{get_mailinglist}
\alias{get_mailinglist}
\title{Retrieve a specific mailing list}
\usage{
get_mailinglist(mailinglist_id)
}
\arguments{
\item{mailinglist_id}{the mailing list id}
}
\value{
A \code{list} with the mailing list information
}
\description{
Retrieve a specific mailing list
}
\examples{
\dontrun{get_mailinglist("ML_1234567890AbCdE")}
}
