% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{get_message}
\alias{get_message}
\title{get_message retrieve a message}
\usage{
get_message(library_id, message_id)
}
\arguments{
\item{library_id}{the library_id}

\item{message_id}{the message_id}
}
\value{
A \code{list}.
}
\description{
get_message retrieve a message
}
