% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_organization}
\alias{get_organization}
\title{get_organization retrieves organization level information}
\usage{
get_organization(organization_id)
}
\arguments{
\item{organization_id}{is the name of the organization}
}
\value{
A list.
}
\description{
get_organization retrieves organization level information
}
\examples{
\dontrun{get_organization("example organization")}

}
