% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_organization_information}
\alias{get_organization_information}
\title{Retrieve general information about an organization}
\usage{
get_organization_information(organization_id)
}
\arguments{
\item{organization_id}{is the name of the organization}
}
\value{
A list.
}
\description{
The organization is also called the brand. It is shown when you hover the mouse over the person icon at the top of your organization's qualtrics.com screen.
}
\note{
To use this call you must be a Qualtrics brand administrator
}
\examples{
\dontrun{get_organization_information("my organization")}

}
