% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-imports.R
\name{get_response_import_progress}
\alias{get_response_import_progress}
\title{start_response_import_url imports a survey based on a loacl survey file}
\usage{
get_response_import_progress(surveyId, progressId)
}
\arguments{
\item{surveyId}{the name of the survey where to import responses}

\item{progressId}{id of the response import}
}
\value{
A \code{list}
}
\description{
start_response_import_url imports a survey based on a loacl survey file
}
\examples{
\dontrun{start_response_import_url("SV_1873930DS2", "11-00-22")}
}
