% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{get_sample}
\alias{get_sample}
\title{Retrieve a sample from a mailing list}
\usage{
get_sample(mailinglist_id, sample_id)
}
\arguments{
\item{mailinglist_id}{the mailing list id}

\item{sample_id}{the contact id}
}
\value{
A \code{list} with all sample contacts and their respective email histories
}
\description{
Retrieve a sample from a mailing list
}
\examples{
\dontrun{get_sample("ML_0HT4q2Ni634kLhH", "PL_8C7RS264INboiMJ")}
}
