% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_sms_distribution}
\alias{get_sms_distribution}
\title{Retrieve an sms survey distribution object}
\usage{
get_sms_distribution(smsdistribution_id, survey_id)
}
\arguments{
\item{smsdistribution_id}{the distribution id}

\item{survey_id}{the distribution id}
}
\value{
A \code{list}
}
\description{
Retrieve an sms survey distribution object
}
\examples{
\dontrun{get_sms_distribution("SMSD_012345678901234", "SV_012345678901234")}
}
