% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{get_survey}
\alias{get_survey}
\title{Retrieve a survey}
\usage{
get_survey(survey_id)
}
\arguments{
\item{survey_id}{the id of survey to retrieve}
}
\value{
A \code{list} of survey elements
}
\description{
Retrieve a survey
}
\examples{
\dontrun{get_survey("SV_012345678901234")}
}
