% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-flows.R
\name{get_survey_flow}
\alias{get_survey_flow}
\title{Retrieve a survey question}
\usage{
get_survey_flow(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A list
}
\description{
Retrieve a survey question
}
\examples{
\dontrun{get_survey_flow("SV_012345678901234")}
}
