% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{get_survey_languages}
\alias{get_survey_languages}
\title{Retrieve the list of languages in which a survey is available}
\usage{
get_survey_languages(survey_id)
}
\arguments{
\item{survey_id}{the id of survey to copy}
}
\value{
A \code{list} of available languages
}
\description{
Retrieve the list of languages in which a survey is available
}
\examples{
\dontrun{get_survey_languages("SV_012345678901234")}
}
