% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-survey.R
\name{get_survey_meta_data}
\alias{get_survey_meta_data}
\title{Retrieve Survey Meta Data}
\usage{
get_survey_meta_data(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A list
}
\description{
Retrieve Survey Meta Data
}
\examples{
\dontrun{get_options("SV_012345678901234")}
}
