% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-options.R
\name{get_survey_options}
\alias{get_survey_options}
\title{Retrieve Survey Options}
\usage{
get_survey_options(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\description{
Retrieve Survey Options
}
\examples{
\dontrun{get_options("SV_012345678901234")}
}
