% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-questions.R
\name{get_survey_question}
\alias{get_survey_question}
\title{Retrieve a survey question}
\usage{
get_survey_question(survey_id, question_id)
}
\arguments{
\item{survey_id}{the survey id}

\item{question_id}{the question id (see list_survey_questions)}
}
\value{
A list
}
\description{
Retrieve a survey question
}
