% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{get_survey_response_counts}
\alias{get_survey_response_counts}
\title{Retrieve a survey}
\usage{
get_survey_response_counts(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A list of responses by response type
}
\description{
Retrieve a survey
}
\examples{
\dontrun{get_survey_response_counts("SV_012345678901234")}
}
