% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{get_survey_responses exports a survey into R
Export survey responses into R and/or save exported file. Currently single exports calls are limited to 1.8 GB. Above that size, we recommend you split your export accross different date ranges}
\usage{
get_survey_responses(surveyId, format = "json", saveDir = NULL,
  filename = NULL, verbose = FALSE, startDate = NULL,
  endDate = NULL, limit = NULL, useLabels = "true",
  seenUnansweredRecode = "false",
  multiSelectSeenUnansweredRecode = "false",
  includeDisplayOrder = "false", formatDecimalAsComma = "false",
  timeZone = NULL, newlineReplacement = "\\n", questionIds = NULL,
  embeddedDataIds = NULL, surveyMetadataIds = NULL,
  compress = "true")
}
\arguments{
\item{surveyId}{the id of survey to copy}

\item{format}{file format json, by default (can be csv or tsv). We don't provide SPSS yet.}

\item{saveDir}{path to a local directory to save the exported file. Default is a temporary file in rds format that is removed as the end of your session. Default name will be the surveyId.}

\item{filename}{specify filename for saving. If NULL, uses the survey id}

\item{verbose}{prints progress bars for download (set to false by default)}

\item{startDate}{only export responses recorded after the specified date (2016-04-01T07:31:43Z) (see \url{https://api.qualtrics.com/docs/dates-and-times})}

\item{endDate}{only export responses before the specified date}

\item{limit}{maximum number of respones exported}

\item{useLabels}{instead of exporting the recode value for the answer choice, export the text of the answer choice (\url{https://www.qualtrics.com/support/survey-platform/survey-module/question-options/recode-values/})}

\item{seenUnansweredRecode}{recode seen but unanswered questions with this value.}

\item{multiSelectSeenUnansweredRecode}{recode seen but unanswered choices for multi-select questions. Default is seenUnansweredRecode value.}

\item{includeDisplayOrder}{include display order information in your export. Useful for surveys with randomization (true or false)}

\item{formatDecimalAsComma}{use a comma as a decimal separator instead of a period (true or false)}

\item{timeZone}{timezone used to determine response date values (default is UTC/GMT). See \url{https://api.qualtrics.com/docs/time-zones}.}

\item{newlineReplacement}{configure the newline delimiter for the export (default is \code{'\n'}.)}

\item{questionIds}{JSON array of Question IDs e.g. (\code{["QID1", "QID3", ... , "QID5"]})}

\item{embeddedDataIds}{export only specified embeddeddata (array of strings)}

\item{surveyMetadataIds}{export only specified survey metadata columns (array of strings)}

\item{compress}{compressed by default (set to FALSE if uncompressed file desired)}
}
\value{
A status code
}
\description{
get_survey_responses exports a survey into R
Export survey responses into R and/or save exported file. Currently single exports calls are limited to 1.8 GB. Above that size, we recommend you split your export accross different date ranges
}
\examples{
\dontrun{get_survey_responses("SV_012345678901234","csv")}
\dontrun{
mysurvey <- getSurvey(surveyID = "SV_012345678901234",
                      startDate = "2017-01-01",
                      endDate = "2017-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose = TRUE)
                      }
}
\seealso{
\url{https://api.qualtrics.com/docs/getting-survey-responses-new} for more documentation
}
