% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{get_survey_responses_new}
\alias{get_survey_responses_new}
\title{Export a survey's responses into R}
\usage{
get_survey_responses_new(survey_id, format = "json", verbose = FALSE,
  saveDir = NULL, filename = NULL, ...)
}
\arguments{
\item{survey_id}{the id of survey to copy}

\item{format}{file format json, by default (can be csv or tsv). We don't provide SPSS yet.}

\item{verbose}{default FALSE}

\item{saveDir}{path to a local directory to save the exported file. Default is a temporary file in rds format that is removed as the end of your session. Default name will be the surveyId.}

\item{filename}{specify filename for saving. If NULL, uses the survey id}

\item{...}{a list of named parameters, see
\url{https://api.qualtrics.com/docs/getting-survey-responses-new} for
parameter names}
}
\value{
A status code
}
\description{
Export survey responses into R and/or save exported file. Currently single
exports calls are limited to 1.8 GB. Above that size, we recommend you split
your export accross different date ranges
}
\examples{
\dontrun{get_survey_responses("SV_012345678901234", "csv")}
\dontrun{get_survey_responses("SV_012345678901234", "csv", filename = "name_export")}
\dontrun{
mysurvey <- getSurvey(surveyID = "surveyId",
                      startDate = "2017-01-01",
                      endDate = "2017-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose = TRUE)
                      }
}
