% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-users.R
\name{get_user}
\alias{get_user}
\title{Retrieve a user account's information}
\usage{
get_user(user_id)
}
\arguments{
\item{user_id}{the user id}
}
\value{
A \code{list} of the different available account parameters.
}
\description{
Retrieve a user account's information
}
\examples{
\dontrun{get_user("UR_012345678912345")}
}
