% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-users.R
\name{get_user_api_token}
\alias{get_user_api_token}
\title{Returns the active api token for an account}
\usage{
get_user_api_token(user_id)
}
\arguments{
\item{user_id}{the user account id}
}
\value{
An api token value
}
\description{
Returns the active api token for an account
}
\examples{
\dontrun{get_user_api_token("UR_012345678912345")}
}
