% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{import_survey}
\alias{import_survey}
\title{Import a survey into Qualtrics}
\usage{
import_survey(upload_name, file_path, file_type = "qsf")
}
\arguments{
\item{upload_name}{the name of the imported survey}

\item{file_path}{path to the file}

\item{file_type}{qsf, txt or doc}
}
\value{
Id of the imported survey
}
\description{
Import a survey into Qualtrics
}
\examples{
\dontrun{import_survey("test_name", "./survey.qsf", 1)}
\dontrun{import_survey("test_name", "./survey.txt", 2)}
}
