% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{import_survey_fromurl}
\alias{import_survey_fromurl}
\title{Import a survey into Qualtrics based on a file URL}
\usage{
import_survey_fromurl(upload_name, file_url, file_type = "qsf")
}
\arguments{
\item{upload_name}{the name of the imported survey}

\item{file_url}{URL to file}

\item{file_type}{qsf, txt or doc}
}
\value{
A status code
}
\description{
Import a survey into Qualtrics based on a file URL
}
\examples{
\dontrun{import_survey("Test", "https://www.example.com/mysurvey.qsf", 1)}
}
