% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{insert_embedded_data}
\alias{insert_embedded_data}
\title{Insert embedded data fields into a survey}
\usage{
insert_embedded_data(survey_id, list_fields)
}
\arguments{
\item{survey_id}{the survey id}

\item{list_fields}{A list of list}
}
\description{
Insert embedded data fields into a survey
}
\examples{
\dontrun{insert_embedded_data("SV_012345678901234", "sss")}
}
