% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{is_success}
\alias{is_success}
\alias{is_success.qualtrics_download}
\title{Success?}
\usage{
is_success(requests, verbose = FALSE)

\method{is_success}{qualtrics_download}(requests, verbose = FALSE)
}
\arguments{
\item{requests}{Requests, as returned by \code{request_downloads}}

\item{verbose}{prints progress bars for download (set to false by default)}
}
\description{
Check which request was successful and which was not.
}
\examples{
requests <- request_downloads(c(1,2))
success <- is_success(requests, TRUE)

}
