% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{list_blocks}
\alias{list_blocks}
\title{List all block ids}
\usage{
list_blocks(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A tibble with blocks' information
}
\description{
This is convenience function that uses the get_survey function to return the list of all
block ids in a survey
}
\examples{
\dontrun{
list_blocks(
"SV_012345678901234", 
)
}
}
