% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{list_contacts}
\alias{list_contacts}
\title{Retrieve all contacts in a mailing list}
\usage{
list_contacts(mailinglist_id)
}
\arguments{
\item{mailinglist_id}{the id of mailing list to delete}
}
\value{
A \code{tibble} of all mailing list contacts
}
\description{
Retrieve all contacts in a mailing list
}
\examples{
\dontrun{list_contacts("ML_1234567890AbCdE")}
}
