% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xm-directory-api.R
\name{list_directory_contacts}
\alias{list_directory_contacts}
\title{List all directory contacts}
\usage{
list_directory_contacts()
}
\value{
A \code{tibble} of all mailing list with properties
}
\description{
List all directory contacts
}
\examples{
\dontrun{list_directory_contacts()}
}
