% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{list_distribution_links}
\alias{list_distribution_links}
\title{List the survey links associated with a distribution}
\usage{
list_distribution_links(distribution_id, survey_id)
}
\arguments{
\item{distribution_id}{the distribution id}

\item{survey_id}{the id of survey}
}
\value{
A \code{tibble}.
}
\description{
Distribution links can only be generated for distribution ids associated
with mailing lists. It is not possible to retrieve survey links for
distributions to only one contact id.
}
\examples{
\dontrun{
 lst_distri <- list_distributions("SV_dg0P0pcZoDYvpNX")
 df <- list_distribution_links("EMD_3rrDZa8AQBnACYl", "SV_dg0P0pcZoDYvpNX")   
}

}
