% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{list_distributionlinks}
\alias{list_distributionlinks}
\title{List the survey links associated with a distribution}
\usage{
list_distributionlinks(distribution_id, survey_id)
}
\arguments{
\item{distribution_id}{the distribution id}

\item{survey_id}{the id of survey}
}
\value{
A \code{tibble}.
}
\description{
Distribution Links can only be generated for distribution ids associated
with mailing lists
}
