% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List all distributions associated to a survey}
\usage{
list_distributions(survey_id)
}
\arguments{
\item{survey_id}{the id of survey}
}
\value{
A \code{list}.
}
\description{
List all distributions associated to a survey
}
