% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-data-privacy.R
\name{list_erasing_requests}
\alias{list_erasing_requests}
\title{list_erasing_requests lists all erasing requests}
\usage{
list_erasing_requests()
}
\value{
A \code{data.frame}.
}
\description{
list_erasing_requests lists all erasing requests
}
\examples{
\dontrun{list_erasing_requests()}
}
