% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{list_event_types}
\alias{list_event_types}
\title{Retrieve the type of activity events}
\usage{
list_event_types()
}
\value{
A \code{tibble}
}
\description{
Retrieve the type of activity events
}
\examples{
\dontrun{list_event_types()}
}
