% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{list_groups}
\alias{list_groups}
\title{List all existing groups for the organization}
\usage{
list_groups()
}
\value{
A \code{tibble} of available groups.
}
\description{
List all existing groups for the organization
}
\examples{
\dontrun{list_groups()}

}
