% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{list_libraries}
\alias{list_libraries}
\title{List all available libraries in the account}
\usage{
list_libraries()
}
\value{
A \code{list}.
}
\description{
List all available libraries in the account
}
\examples{
\dontrun{
 list_libraries()
}
}
