% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{list_library_blocks}
\alias{list_library_blocks}
\title{Lists the library blocks for the caller of this API.}
\usage{
list_library_blocks(library_id)
}
\arguments{
\item{library_id}{the library id (see 'list_libraries')}
}
\value{
A \code{tibble}
}
\description{
Lists the library blocks for the caller of this API.
}
\examples{
\dontrun{
 list_library_blocks("UR_0NXtl92JJWqfWcJ")
}
}
