% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{list_library_surveys}
\alias{list_library_surveys}
\title{Lists the library surveys for the caller of this API.}
\usage{
list_library_surveys(library_id)
}
\arguments{
\item{library_id}{the library id (see 'list_libraries')}
}
\value{
A tibble
}
\description{
Lists the library surveys for the caller of this API.
}
\examples{
\dontrun{
 list_library_surveys("UR_0NXtl92JJWqfWcJ")
}
}
