% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{list_mailinglists}
\alias{list_mailinglists}
\title{Retrieve all mailing lists from account}
\usage{
list_mailinglists()
}
\value{
A \code{tibble} of all mailing list with properties
}
\description{
Retrieve all mailing lists from account
}
\examples{
\dontrun{list_mailinglists()}
}
