% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{list_samples}
\alias{list_samples}
\title{List samples associated with a mailing list}
\usage{
list_samples(mailinglist_id)
}
\arguments{
\item{mailinglist_id}{the id of mailing list to delete}
}
\value{
A \code{tibble} of all samnple ids and names
}
\description{
List samples associated with a mailing list
}
\examples{
\dontrun{list_samples("ML_0HT4q2Ni634kLhH")}
}
