% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{list_sms_distributions}
\alias{list_sms_distributions}
\title{List all sms distributions associated to a survey}
\usage{
list_sms_distributions(survey_id)
}
\arguments{
\item{survey_id}{the distribution id}
}
\value{
A \code{list}
}
\description{
List all sms distributions associated to a survey
}
\examples{
\dontrun{list_sms_distributions("SV_012345678901234")}
}
