% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{list_surveys}
\alias{list_surveys}
\title{Retrieve the list of surveys for a given user account}
\usage{
list_surveys()
}
\value{
A \code{tibble}.
}
\description{
Retrieve the list of surveys for a given user account
}
\examples{
\dontrun{list_surveys()}
}
