% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{list_users}
\alias{list_users}
\title{Retrieve the list of all user accounts}
\usage{
list_users()
}
\value{
A \code{tibble}.
}
\description{
Retrieve the list of all user accounts
}
\examples{
\dontrun{list_users()}
}
