% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qualtrics_auth}
\alias{qualtrics_auth}
\alias{load_oauth}
\alias{refresh_oauth}
\alias{set_qualtrics_opts}
\title{Authenticate}
\usage{
qualtrics_auth(id, secret, data_center = NULL, cache = TRUE)

load_oauth()

refresh_oauth()

set_qualtrics_opts(api_token, data_center = NULL)
}
\arguments{
\item{id, secret}{Your client id and secret.}

\item{data_center}{Your data center id, if \code{NULL} then the functions attempts to
get it from \code{\link{set_qualtrics_opts}}.}

\item{cache}{Whether to set the token as option (\code{\link{set_qualtrics_opts}})
and cache the results to file.}

\item{api_token}{Your client id and generated token.}
}
\value{
if \code{cache} is \code{FALSE} then returns an object of class \code{quatrics_token}.
}
\description{
Generate a Qualtrics token valid for an hour.
}
\details{
Obtains a bearer token valid for an hour and saved it to the working
directory in a hidden file named \code{qualtrics-oauth}. Note that this token
will be automatically loaded in future sessions, and refreshed if needed.
}
\section{Functions}{

\itemize{
  \item{\code{qualtrics_auth}: Authenticate.}
  \item{\code{load_oauth}: Load locally saved token.}
  \item{\code{refresh_oauth}: Refresh locally saved token.}
  \item{\code{set_qualtrics_opts}: Set \code{token} and \code{data_center} options.}
}
}

\examples{
\dontrun{
  # OAUth 2.0
  qualtrics_auth("xXxX0x0X0xxX0", "xXxX0x0X0xxX0", "data.center")

  # Set token and data_center options
  set_qualtrics_opts("xXxX0x0X0xxX0xXxX0x0X0xxX0", "myCenter.eu")
}

}
