% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_opts}
\alias{.get_opts}
\alias{.get_url}
\alias{.get_data_center}
\title{Convenience function
Fetches option.}
\usage{
.get_opts(opts = NULL)

.get_url()

.get_data_center()
}
\arguments{
\item{opts}{Name of option.}
}
\description{
Convenience function
Fetches option.
}
\details{
\code{.get_url} and \code{.get_data_center} are simple convenience
as these are used often.
}
