% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{remove_user_group}
\alias{remove_user_group}
\title{Remove a user from a given group}
\usage{
remove_user_group(group_id, user_id)
}
\arguments{
\item{group_id}{group id}

\item{user_id}{id of the user account to remove from the group}
}
\value{
A status code
}
\description{
Remove a user from a given group
}
\examples{
\dontrun{remove_user_group("GR_1234567890AbCdE", "UR_1234567890AbCdE")}
}
