% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{send_email_tomailinglist}
\alias{send_email_tomailinglist}
\title{Send Email to Mailing List}
\usage{
send_email_tomailinglist(
  fromEmail,
  fromName,
  replyToEmail,
  subject,
  messageText,
  mailingListId,
  sendDate
)
}
\arguments{
\item{fromEmail}{To generate transaction distribution links, value must be "CreateTransactionBatchDistribution"}

\item{fromName}{the distribution id}

\item{replyToEmail}{the distribution id}

\item{subject}{the distribution id}

\item{messageText}{the id of survey}

\item{mailingListId}{the id of survey}

\item{sendDate}{the id of survey}
}
\value{
The email distribution id
}
\description{
Create a new distribution with a custom message
}
