% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{share_survey}
\alias{share_survey}
\title{Share a survey with another account}
\usage{
share_survey(survey_id, user_id, surv_permissions)
}
\arguments{
\item{survey_id}{the survey id}

\item{user_id}{the user id of the account with whom you are sharing the survey}

\item{surv_permissions}{array of numbers for permissions activation}
}
\value{
A status execution code
}
\description{
Share a survey with another account
}
\examples{
\dontrun{share_survey("SV_012345678901234", "UR_012345678912345", 15)}
\dontrun{share_survey("SV_012345678901234", "UR_012345678912345", c(15, 23))}
}
