% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-imports.R
\name{start_response_import}
\alias{start_response_import}
\title{start_response_import imports a survey based on a loacl survey file}
\usage{
start_response_import(surveyId, filePath, format = "CSV",
  charset = "UTF-8")
}
\arguments{
\item{surveyId}{the name of the survey where to import responses}

\item{filePath}{The URL of the import file. Only HTTP and HTTPS URLs are supported. The largest file you can import is limited to 5 MB.}

\item{format}{default CSV Allowed values are CSV or TSV}

\item{charset}{Allowed values are UTF-8 or UTF-16}
}
\value{
A status code
}
\description{
start_response_import imports a survey based on a loacl survey file
}
\examples{
\dontrun{start_response_import("SV_012345678912345", "./data.csv", "CSV")}
}
