% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-imports.R
\name{start_response_import_url}
\alias{start_response_import_url}
\title{start_response_import_url imports a survey based on a loacl survey file}
\usage{
start_response_import_url(surveyId, fileUrl, format = "CSV", charset = "UTF-8")
}
\arguments{
\item{surveyId}{the name of the survey where to import responses}

\item{fileUrl}{The URL of the import file. Only HTTP and HTTPS URLs are supported. The largest file you can import is limited to 5 MB.}

\item{format}{default CSV Allowed values are CSV or TSV}

\item{charset}{Allowed values are UTF-8 or UTF-16}
}
\value{
A \code{list}
}
\description{
start_response_import_url imports a survey based on a loacl survey file
}
\examples{
\dontrun{start_response_import_url("SV_012345678912345", "./data.csv", "CSV")}
}
