% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{update_block}
\alias{update_block}
\title{Update a block}
\usage{
update_block(
  survey_id,
  block_id,
  type = "Standard",
  description = "New Block",
  id = NULL,
  options,
  elements
)
}
\arguments{
\item{survey_id}{the survey id}

\item{block_id}{the block id}

\item{type}{block type}

\item{description}{block description}

\item{id}{a unique block id identifier (should be of form BL_FakeBlockID1234)}

\item{options}{list of block options}

\item{elements}{list of block elements}
}
\value{
A list
}
\description{
You can update an existing block, including the blocks meta information and the the block's elements. 
If you update the block elements and you add for example a new question, you must make sure the question
exisists in the survey (see list_questions and create_question).
}
\examples{
\dontrun{
update_block("SV_012345678901234", "BL_723z35LY23KCZ4p")
}
}
