% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{update_contact}
\alias{update_contact}
\title{Updates a contact's information in a mailing list}
\usage{
update_contact(mailinglist_id, contact_id, contact_options)
}
\arguments{
\item{mailinglist_id}{ID of the mailing list in which to create the contact}

\item{contact_id}{ID of the contact to update (see list_contacts)}

\item{contact_options}{list of new contact information}
}
\value{
The status code
}
\description{
update a contact's information in a mailing list. You can pass a named
list of parameters to define your contact. All parameters are optional but empty parameters
cannot be passed. Parameters are: firstName, lastName, email, externalDataRef, language,
unsubscribed and embeddedData. The embeddedData element must a string representation of a JSON object.
}
\examples{
\dontrun{
updated_contact <- list("firstName" = "James William")
update_contact("ML_0HT4q2Ni634kLhH", "MLRP_eaI7CCX9bxAbRt3", updated_contact)}
}
