% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-flows.R
\name{update_flow_element}
\alias{update_flow_element}
\title{Update a specific survey flow element}
\usage{
update_flow_element(survey_id, flow_id, new_flow_id, flow_type)
}
\arguments{
\item{survey_id}{the survey id}

\item{flow_id}{the flow element id}

\item{new_flow_id}{the new flow id}

\item{flow_type}{the flow element type (see docs for available types)}
}
\value{
A list
}
\description{
This call allows only limited action on a flow element. It allows only to update the
type of a flow element as well as its id. Such functionalities are useful only if you
have a detailing understanding of the whole structure of Qualtrics surveys, which is best
understood spending time understanding exported qsf files.
}
