% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{update_group}
\alias{update_group}
\title{Update group's definition parameters}
\usage{
update_group(group_id, type = NULL, name = NULL, division_id = NULL)
}
\arguments{
\item{group_id}{group id, see \code{list_groups()}}

\item{type}{the new group type}

\item{name}{the new group name}

\item{division_id}{default NULL but specific division id can be specified to
automatically assign the division properties}
}
\value{
A status execution code
}
\description{
Currently this is a GAP as you cannot create group types with an API nor
look them up. But you need that to create a Group.
}
\examples{
\dontrun{update_group("GR_1234567890AbCdE", name = "New Name")}
}
