% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-core-contacts.R
\name{update_mailinglist}
\alias{update_mailinglist}
\title{Update an existing mailing list}
\usage{
update_mailinglist(mailinglist_id, library_id = NULL, name = NULL,
  category = NULL)
}
\arguments{
\item{mailinglist_id}{the library_id}

\item{library_id}{the library_id if moved to a different library id}

\item{name}{Name of the new mailing list}

\item{category}{Category in which to create the new mailing list}
}
\value{
A \code{list}.
}
\description{
Update an existing mailing list
}
\examples{
\dontrun{update_mailinglist("ML_1234567890AbCdE", name = "New Name")}
}
