% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-options.R
\name{update_options}
\alias{update_options}
\title{Update Survey Options}
\usage{
update_options(survey_id, options)
}
\arguments{
\item{survey_id}{the survey id}

\item{options}{list of named option defintions}
}
\description{
All available options can be found at \url{https://api.qualtrics.com/reference#update-options}
}
\examples{
\dontrun{update_options("SV_012345678901234", "sss")}
}
