% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{update_response}
\alias{update_response}
\title{Update the value of an embedded data field in a survey response}
\usage{
update_response(surveyId, responseId, embeddedData,
  resetRecordedDate = TRUE)
}
\arguments{
\item{surveyId}{the id of survey to copy}

\item{responseId}{ID of the response to be deleted}

\item{embeddedData}{JSON object representing the embedded data fields to set.}

\item{resetRecordedDate}{Default: true - Sets the recorded date to the current
time. If false, the recorded date will be incremented by one millisecond.}
}
\value{
A \code{status}.
}
\description{
Update the value of an embedded data field in a survey response
}
\examples{
\dontrun{update_response("SV_012345678912345", "R_012345678912345",
{"EDField": "EDValue"}, FALSE)}
}
