% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{update_survey}
\alias{update_survey}
\title{Update a surve's name, status or expiration date}
\usage{
update_survey(survey_id, status, new_name = NULL,
  expiration_dates = NULL)
}
\arguments{
\item{survey_id}{the id of survey to copy}

\item{status}{is the activity status of the survey (true or false)}

\item{new_name}{is the edited name of the survey}

\item{expiration_dates}{an array with startDate and endDate (format "2016-01-01T01:00:00Z")}
}
\value{
A execution status code
}
\description{
Update a surve's name, status or expiration date
}
\examples{
\dontrun{update_survey("SV_012345678901234", true)}
\dontrun{update_survey("SV_012345678901234", true, "new_name", c("2016-01-01T01:00:00Z","2016-03-01T01:00:00Z"))}
}
