% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-languages.R
\name{update_survey_translations}
\alias{update_survey_translations}
\title{Update survey translation}
\usage{
update_survey_translations(survey_id, language_code, survey_translation)
}
\arguments{
\item{survey_id}{the survey id}

\item{language_code}{the language code}

\item{survey_translation}{list with all translated survey fields (see \code{get_survey_translations})}
}
\value{
A \code{status}.
}
\description{
Update survey translation
}
\details{
Update a surveys translation by provind a list of survey elements' translations for a
given language.
}
\examples{
\dontrun{
# retrieve translation from a survey
srv_transl_pt <- get_survey_translations("SV_dnEGNjwrSTQXxiZ", "PT")
# create pt translation in other survey
update_survey_languages("SV_6fj3YgWt6ocXL1A", "PT")
# update the PT translation
update_survey_translations(id, "PT", srv_transl_pt)
}
}
