% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{update_user}
\alias{update_user}
\title{Update a user's account parameters}
\usage{
update_user(user_id, ...)
}
\arguments{
\item{user_id}{id of the user account to update}

\item{...}{a named list of account parameters to update}
}
\value{
A status execution code
}
\description{
Update a user's account parameters
}
\section{Account parameters}{

user settings that can be possed are user_name, first_name, last_name,
user_type, language, time_zone, division_id, account_expiration_date,
status and permissions
}

\examples{
\dontrun{update_user("UR_012345678912345", c("status" = "Enabled", "language" = "EN"))}
}
