% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic-library.R
\name{upload_graphic}
\alias{upload_graphic}
\title{upload_graphic imports a survey based on a loacl survey file}
\usage{
upload_graphic(libraryId, filename, filePath, type, folder = NULL)
}
\arguments{
\item{libraryId}{the name of the survey where to import responses}

\item{filename}{Name of the graphic}

\item{filePath}{Path to the file to import}

\item{type}{image/jpeg, image/gif, image/png}

\item{folder}{The folder where the graphic will be located}
}
\value{
the id of the uploaded graphic
}
\description{
upload_graphic imports a survey based on a loacl survey file
}
\examples{
\dontrun{upload_graphic("SV_012345678912345", "./data.csv", "CSV")}
}
