% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic-library.R
\name{upload_graphic_fromurl}
\alias{upload_graphic_fromurl}
\title{upload_graphic_fromurl imports a survey based on a loacl survey file}
\usage{
upload_graphic_fromurl(libraryId, name, fileUrl, contentType, folder = NULL)
}
\arguments{
\item{libraryId}{the name of the survey where to import responses}

\item{name}{Name of the graphic}

\item{fileUrl}{URL to the file to import}

\item{contentType}{image/jpeg, image/gif, image/png}

\item{folder}{The folder where the graphic will be located}
}
\value{
the id of the uploaded graphic
}
\description{
upload_graphic_fromurl imports a survey based on a loacl survey file
}
\examples{
\dontrun{upload_graphic("SV_012345678912345", "./data.csv", "CSV")}
}
