% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{who_am_i}
\alias{who_am_i}
\title{Determine the user ID and other user information associated with an Qualtrics
API token or a Qualtrics OAuth access token.}
\usage{
who_am_i()
}
\value{
A \code{list} of related account information.
}
\description{
Determine the user ID and other user information associated with an Qualtrics
API token or a Qualtrics OAuth access token.
}
\examples{
\dontrun{who_am_i()}
}
