% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-groups.R
\name{add_user_group}
\alias{add_user_group}
\title{Add a new user to a given group}
\usage{
add_user_group(group_id, user_id)
}
\arguments{
\item{group_id}{group id}

\item{user_id}{id of the user account to add to a group}
}
\value{
A status code
}
\description{
Add a new user to a given group
}
\examples{
\dontrun{add_user_group("GR_1234567890AbCdE", "UR_1234567890AbCdE")}
}
