% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{bulk-download}
\alias{bulk-download}
\alias{request_downloads}
\alias{download_requested}
\alias{download_requested.qualtrics_download}
\title{Request downloads}
\usage{
request_downloads(surveyIds, format = "json", ...)

download_requested(requests, format = "json", saveDir = NULL)

\method{download_requested}{qualtrics_download}(requests,
  format = "json", saveDir = NULL, verbose = FALSE)
}
\arguments{
\item{surveyIds}{A vector of survey ids.}

\item{format}{file format json, by default (can be csv or tsv). We don't provide SPSS yet.}

\item{...}{a vector of named parameters, see \url{https://api.qualtrics.com/reference} *Create Response Export* for parameter names}

\item{requests}{the list output of request_downloads}

\item{saveDir}{the output dir to save the data}

\item{verbose}{print progress for heavy downloads (default is FALSE)}

\item{format}{the output file format}
}
\value{
A \code{request_downloads} returns object of class \code{qualtrics_download} while
\code{download_requested} returns a \code{list}.
}
\description{
Request survey response downloads.
}
\details{
The \href{https://www.qualtrics.com}{Qualtrics} API downloads survey responses in several steps.
First a download is \emph{requested}. Qualtrics prepares the file and \code{\link{download_requested}}
can be triggered to download the file when it is ready. When downloading responses from a large nuimber of
surveys, it is more efficient to request all downloads if a first place, check that the downloads are ready
and then retrieve all the files. The performances will headvily depend on the number
of surveys and the amount of data the user ultimately wants to obtain.
}
\section{Functions}{

\itemize{
  \item{\code{request_downloads}: Request download (prepares files).}
  \item{\code{download_requested}: Download requested files.}
}
}

\examples{
\dontrun{
requests <- request_downloads(c(1,2)))
data <- download_requested(requests)}

}
