% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{create_distribution}
\alias{create_distribution}
\title{Create email distribution or generate distribution links}
\usage{
create_distribution(
  survey_id,
  library_id,
  message_id,
  mailing_list_id,
  from_name,
  reply_to_email,
  from_email,
  subject,
  type,
  send_date = Sys.Date() + 1
)
}
\arguments{
\item{survey_id}{the survey id}

\item{library_id}{the library id where the message is saved (see 'list_libraries')}

\item{message_id}{id of the message item}

\item{mailing_list_id}{mailing list is, or set transaction_batch_id}

\item{from_name}{appearing sender name, default is Qualtrics}

\item{reply_to_email}{optional, reply email, default is \email{noreply@qualtrics.com}}

\item{from_email}{sender email, default is \email{noreply@qualtrics.com}}

\item{subject}{email subject, default is "Survey Distribution"}

\item{type}{Allowed values: Individual, Multiple, Anonymous}

\item{send_date}{date for distribution to be sent}

\item{transaction_batch_id}{transaction batch, or set mailing_list_id}

\item{contact_id}{optional, optional contact lookup ID for individual distribution}

\item{expiration_date}{expiration date}
}
\value{
The created distribution id
}
\description{
Using this API call, you can either schedule an email distribution
using the Qualtrics email server, or generate distribution links
that you can then use to share the link to your survey using your own
email solution.
}
\details{
In order to create a new email distribution using the API, you need
to have the body of your email already created as a message in your
Qualtrics message library (see 'list_messages' and 'create_message' 
functions).
}
