% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-privacy.R
\name{create_erasure_request}
\alias{create_erasure_request}
\title{Create an erasing request}
\usage{
create_erasure_request(list_emails, search_only = TRUE)
}
\arguments{
\item{list_emails}{A vector of emails.}

\item{search_only}{Set to \code{TRUE} to perform a dryrun (set to FALSE to erase)}
}
\value{
A the request id
}
\description{
When processed, the request will search for any entities within Qualtrics that
match any of the supplied emails; these entities include
\itemize{
\item Users
\item MailingList & XM Directory contacts
\item People
For any entity discovered the service will erase any responses to surveys
they have submitted, along with any tickets that may have originated from
their responses before erasing the entity itself.
}
}
\examples{
\dontrun{create_erasure_request("ddd@hotmail.com")}
\dontrun{create_erasure_request(c("person1@themail.com", "person1@theemail.com"), FALSE)}

}
