% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-blocks.R
\name{get_block}
\alias{get_block}
\title{Retrieve a given survey block}
\usage{
get_block(survey_id, block_id)
}
\arguments{
\item{survey_id}{the survey id}

\item{block_id}{the block id}
}
\value{
A list
}
\description{
Retrieve a given survey block
}
\examples{
\dontrun{get_block("SV_012345678901234","BL_723z35LY23KCZ4p")}
}
