% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-divisions.R
\name{get_division}
\alias{get_division}
\title{Retrieve a given division}
\usage{
get_division(division_id)
}
\arguments{
\item{division_id}{group id}
}
\value{
A \code{list} of division definition parameters
}
\description{
Retrieve a given division
}
\examples{
\dontrun{get_division("DV_1234567890AbCdE")}
}
