% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-data-privacy.R
\name{get_erasure_request}
\alias{get_erasure_request}
\title{Retrieve a given erasure request}
\usage{
get_erasure_request(request_id)
}
\arguments{
\item{request_id}{the id of the past erasing request}
}
\value{
A \code{list}.
}
\description{
Retrieve a given erasure request
}
\examples{
\dontrun{get_erasing_request("request_id")}
}
