% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{get_group}
\alias{get_group}
\title{Retrieve group information}
\usage{
get_group(group_id)
}
\arguments{
\item{group_id}{group id}
}
\value{
A \code{list} of group definition parameters
}
\description{
Retrieve group information
}
\examples{
\dontrun{update_group("GR_1234567890AbCdE")}
}
