% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{get_message}
\alias{get_message}
\title{Retrieve a message from the library}
\usage{
get_message(library_id, message_id)
}
\arguments{
\item{library_id}{The library id can be one of the following IDs: User ID, Group ID, ThreeSixty ID (see 'list_libraries')}

\item{message_id}{the message_id}
}
\value{
A \code{list}.
}
\description{
The call returns a list with the message category, the message description
and the message in the various translated languages.
}
\examples{
\dontrun{
 get_message("UR_0NXtl92JJWqfWcJ", "MS_0fddN2xI3J0nGQZ")
}

}
