% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-organizations.R
\name{get_organization}
\alias{get_organization}
\title{get_organization retrieves organization level information}
\usage{
get_organization()
}
\value{
A \code{data.frame}.
}
\description{
get_organization retrieves organization level information
}
\examples{
\dontrun{get_organization()}
}
