% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-flows.R
\name{get_survey_flow}
\alias{get_survey_flow}
\title{Retrieve the flow structure of a survey}
\usage{
get_survey_flow(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A list
}
\description{
This call will return the complete flow strcuture of your survey as a
nested list of flow elements reflecting all layers of the logic built into the
survey. You can extract block ids from the response to use with the get_block
function.
}
\examples{
\dontrun{
get_survey_flow("SV_012345678901234")
}
}
