% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-languages.R
\name{get_survey_languages}
\alias{get_survey_languages}
\title{Retrieve the list of languages in which a survey is available}
\usage{
get_survey_languages(survey_id)
}
\arguments{
\item{survey_id}{the survey id}
}
\value{
A \code{list} of available languages
}
\description{
Retrieve the list of languages in which a survey is available
}
\examples{
\dontrun{get_survey_languages("SV_012345678901234")}

}
