% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-response-exports.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Export a survey's responses into R}
\usage{
get_survey_responses(survey_id, format = "json", verbose = FALSE,
  saveDir = NULL, filename = NULL, ...)
}
\arguments{
\item{survey_id}{the id of survey to copy}

\item{format}{file format json, by default (can be csv or tsv). We don't provide SPSS yet.}

\item{verbose}{default FALSE}

\item{saveDir}{path to a local directory to save the exported file. Default is a temporary file in rds format that is removed at the end of your session. Default name will be the surveyId.}

\item{filename}{specify filename for saving. If NULL, uses the survey id}

\item{...}{a vector of named parameters, see \url{https://api.qualtrics.com/reference} *Create Response Export* for parameter names}
}
\value{
A status code
}
\description{
Export survey responses into R and/or save exported file. Currently single
exports calls are limited to 1.8 GB. Above that size, we recommend you split
your export accross different date ranges. When exporting a large number of
surveys successively, we recommend using the more efficient \code{request_downloads} and
\code{download_requested} functions.
}
\examples{
\dontrun{get_survey_responses("SV_012345678901234", "csv")}
\dontrun{get_survey_responses("SV_012345678901234", format = "csv", verbose = TRUE, saveDir = "./", filename = "name_export", useLabels = TRUE, limit = 10)}

}
