% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-languages.R
\name{get_survey_translations}
\alias{get_survey_translations}
\title{get_survey_translations retrieves the full translation}
\usage{
get_survey_translations(survey_id, language_code)
}
\arguments{
\item{survey_id}{the survey id}

\item{language_code}{is the code for the language of translation}
}
\value{
A \code{list} of survey elements in requested language
}
\description{
get_survey_translations retrieves the full translation
}
\examples{
\dontrun{get_survey_translations("SV_012345678901234", "EN")}
}
