% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-data-privacy.R
\name{list_erasure_requests}
\alias{list_erasure_requests}
\title{List all erasure request for a brand}
\usage{
list_erasure_requests()
}
\value{
A \code{data.frame}.
}
\description{
List all erasure request for a brand
}
\examples{
\dontrun{list_erasing_requests()}

}
