% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{list_messages}
\alias{list_messages}
\title{list_messages retrieves a full survey object}
\usage{
list_messages(library_id)
}
\arguments{
\item{library_id}{the library_id}
}
\value{
A \code{tibble}.
}
\description{
list_messages retrieves a full survey object
}
