% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{remove_user_group}
\alias{remove_user_group}
\title{Remove a user from a given group}
\usage{
remove_user_group(group_id, user_id)
}
\arguments{
\item{group_id}{the id of the group, see \code{list_groups()}}

\item{user_id}{the user account id, see \code{list_users()}}
}
\value{
A status code
}
\description{
Remove a user from a given group
}
\examples{
\dontrun{remove_user_group("GR_1234567890AbCdE", "UR_1234567890AbCdE")}
}
