% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{send_email_tomailinglist}
\alias{send_email_tomailinglist}
\title{Send Email to Mailing List}
\usage{
send_email_tomailinglist(
  mailingListId,
  sendDate = paste0(Sys.Date() + 1, "T00:00:00Z"),
  subject = "Take this survey",
  messageText = "Hello, take this survey",
  fromEmail = "noreply@qualtrics.com",
  fromName = "Qualtrics",
  replyToEmail = "noreply@qualtrics.com"
)
}
\arguments{
\item{mailingListId}{the id of survey}

\item{sendDate}{the date to send the email (default is +1 day)}

\item{subject}{the email subject}

\item{messageText}{email text body}

\item{fromEmail}{sender email, default is \email{noreply@qualtrics.com}}

\item{fromName}{the send name default is Qualtrics}

\item{replyToEmail}{reply email, default is \email{noreply@qualtrics.com}}
}
\value{
The email distribution id
}
\description{
You can use this call to send an email to a mailing list without
having to create the email message in Qualtrics first.
}
\examples{
\dontrun{send_email_tomailinglist(mail_id, "2020-04-13T03:26:00Z")}

}
