% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-divisions.R
\name{update_division}
\alias{update_division}
\title{Update a division's settings}
\usage{
update_division(division_id, name = NULL, status = NULL,
  permissions = NULL)
}
\arguments{
\item{division_id}{id of the division to update}

\item{name}{New division name}

\item{status}{New status of the division, can be "Active", "Disabled"}

\item{permissions}{JSON object with permissions to update}
}
\value{
id of created message
}
\description{
Update a division's settings
}
\examples{
\dontrun{update_division("DV_1234567890AbCdE")}
}
