% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-groups.R
\name{update_group}
\alias{update_group}
\title{Update group's definition parameters}
\usage{
update_group(group_id, type = NULL, name = NULL, division_id = NULL)
}
\arguments{
\item{group_id}{id of group to update}

\item{type}{new group type}

\item{name}{new name of the group to update}

\item{division_id}{default NULL but specific division id can be specified to
automatically assign the division properties}
}
\value{
A status execution code
}
\description{
Update group's definition parameters
}
\examples{
\dontrun{update_group("GR_1234567890AbCdE", name = "New Name")}
}
