% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-library.R
\name{update_message}
\alias{update_message}
\title{update_message creates a message}
\usage{
update_message(description, messages, library_id, message_id)
}
\arguments{
\item{description}{defines a description for the message}

\item{messages}{A JSON object mapping language codes to messages. Existing languages will be replaced; others will be added}

\item{library_id}{The Library ID can be one of the following IDs: User ID, Group ID, ThreeSixty ID}

\item{message_id}{Message ID}
}
\value{
A \code{tibble}.
}
\description{
update_message creates a message
}
