% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-flows.R
\name{update_survey_flow}
\alias{update_survey_flow}
\title{Update the survey flow}
\usage{
update_survey_flow(survey_id, flow_id, flow_type, flow_object, flow_properties)
}
\arguments{
\item{survey_id}{the survey id}

\item{flow_id}{a string the flow element id}

\item{flow_type}{a string  the flow element type (see docs for available types)}

\item{flow_object}{an object the root flow strcuture}

\item{flow_properties}{an object the flow element properties}
}
\value{
A list
}
\description{
This call allows you to modify the existing flow strcuture of a survey. Performing this
operation through the API requires you understand the details of the full structure of a
Qualtrics survey since creating new flow elements, e..g, will require that these elements
refere to survey elements, like Blocks, that exist, i.e. were previously created through
other calls. For a detailed example, see the official package documentation.
}
