% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-definitions-languages.R
\name{update_survey_translations}
\alias{update_survey_translations}
\title{Update the survey translation for a given language}
\usage{
update_survey_translations(survey_id, language_code, survey_field)
}
\arguments{
\item{survey_id}{the id of survey to copy}

\item{language_code}{the language code}

\item{survey_field}{need to be inferred from retrieving a translation first}
}
\value{
A \code{status}.
}
\description{
Update the survey translation for a given language
}
\details{
To understand which the field names to provide as translated input, the
most simple approach is to run the get_survey_translations function for the
corresponding language and built the list of fields the response.
}
\examples{
\dontrun{update_survey_translations(id, "EN", list("QID1_QuestionText" = "Uno"))}
}
